+++ packages/@headlessng-primitives/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
+- added 'disabled' package to read and manage disabled state
 - added 'focus' package to read and manage focus
 
 ### Fixed